+++ inst/htmlwidgets/myIO/src/Chart.js
@@ -22,6 +22,7 @@ import { FacetController } from "./layout/facet-controller.js";
 import { applyARIA } from "./a11y/aria.js";
 import { KeyboardNavigator } from "./a11y/keyboard-nav.js";
 import { DataTableFallback } from "./a11y/data-table.js";
+import { initializeKeyframes, destroyKeyframes } from "./interactions/keyframes.js";
 
 const MIN_CHART_WIDTH = 280;
 const RESIZE_DEBOUNCE_MS = 100;
@@ -190,6 +191,9 @@ export class myIOchart {
       this.dataTable.initialize();
       applyARIA(this);
     }
+    initializeKeyframes(this);
+    this.derived.currentLayers = this.config.layers;
+    this.syncLegacyAliases();
     this.captureLegacyAliases();
     if (this.derived.currentLayers.length > 0) {
       this.setClipPath(this.derived.currentLayers[0].type);
@@ -523,6 +527,7 @@ export class myIOchart {
 
   destroy() {
     this.emit("destroy", {});
+    destroyKeyframes(this);
     clearTimeout(this.runtime && this.runtime.resizeTimer);
     clearTimeout(this.runtime && this.runtime.tooltipHideTimer);
     if (this.facetController) {